server/requirements.txt
# Core LangGraph and LangChain dependencies
langgraph
langchain-core
langchain-openai
langchain-google-genai

# Google AI dependencies
google-generativeai

# FastAPI and web server dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic-settings>=2.0.0

# Pydantic for data validation
pydantic>=2.0.0

# Environment management
python-dotenv>=1.0.0

# Type hints support
typing-extensions>=4.5.0

# Optional: for development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0